volterra-cgpo: float als/ncl, CGPO_LAMBDA (default 1), uncapped i64 max_p_iters, S0=√(-A/C), gauge-fixed pressure output
- Params::new takes f64 als/ncl (fractional sweep) + lambda; map ζ=K/als², C=K/ncl²
- runner reads CGPO_LAMBDA (code-truth λ=1), CGPO_MAX_P_ITERS as i64 (-1 = uncapped)
- S0 = √(-A/C) per flow-solver.py (was hardcoded √2; IC-only, now faithful)
- cgpo_fd writes gauge-fixed p/ snapshots (fixes pure-Neumann pressure drift)

+++ volterra-cgpo/tests/step.rs
@@ -34,7 +34,7 @@ const A_LANDAU: f64 = -K_ELASTIC / 81.0;
 const C_LANDAU: f64 = K_ELASTIC / 81.0;
 const S0: f64 = std::f64::consts::SQRT_2;
 const DT: f64 = 1e-4;
-const MAX_P_ITERS: usize = 20;
+const MAX_P_ITERS: i64 = 20;
 const P_TARGET_REL_CHANGE: f64 = 1e-4;
 
 // ── helpers ───────────────────────────────────────────────────────────────────
